Eliminación de contraseña en conexión a BBDD

+++ src/Service/GenericService.java
@@ -8,13 +8,14 @@ import java.util.List;
 /**
  *
  * @author Daniela Nahir Romero
+ * @param <T>
  */
 public interface GenericService<T> {
     //interfaz generica que define las operaciones basicas que deben implementar todos los servicios
     void insertar (T entidad) throws Exception;
     void actualizar (T entidad) throws Exception;
-    void eliminar (int id_dom)throws Exception;
-    T getByID (int id) throws Exception;
+    void eliminar (long id_dom)throws Exception;
+    T getByID (long id) throws Exception;
     List<T> getAll() throws Exception;
          
 }
